feat: optimize math class size and dependencies
