+++ src/examples/gettingstarted/4_lambertCube/index.ts
@@ -3,6 +3,11 @@ import { fetchImage } from "../../../lib/io/loaders/Image";
 import { ShaderMaterial } from "../../../lib/materials/ShaderMaterial";
 import { Euler } from "../../../lib/math/Euler";
 import { Matrix4 } from "../../../lib/math/Matrix4";
+import {
+  makeMatrix4Perspective,
+  makeMatrix4RotationFromEuler,
+  makeMatrix4Translation,
+} from "../../../lib/math/Matrix4.Functions";
 import { Vector3 } from "../../../lib/math/Vector3";
 import { BufferGeometry } from "../../../lib/renderers/webgl2/buffers/BufferGeometry";
 import { DepthTestFunc, DepthTestState } from "../../../lib/renderers/webgl2/DepthTestState";
@@ -23,11 +28,10 @@ async function init(): Promise<null> {
   document.body.appendChild(canvasFramebuffer.canvas);
 
   const program = new Program(context, material);
-  const texImage2D = new TexImage2D(context, texture);
   const uniforms = {
     localToWorld: new Matrix4(),
-    worldToView: new Matrix4().makeTranslation(new Vector3(0, 0, -1)),
-    viewToScreen: new Matrix4().makePerspective(-0.25, 0.25, 0.25, -0.25, 0.1, 4.0),
+    worldToView: makeMatrix4Translation(new Matrix4(), new Vector3(0, 0, -1)),
+    viewToScreen: makeMatrix4Perspective(new Matrix4(), -0.25, 0.25, 0.25, -0.25, 0.1, 4.0),
     viewLightPosition: new Vector3(0, 0, 0),
     map: new TexImage2D(context, texture),
   };
@@ -38,7 +42,10 @@ async function init(): Promise<null> {
     requestAnimationFrame(animate);
 
     const now = Date.now();
-    uniforms.localToWorld.makeRotationFromEuler(new Euler(now * 0.001, now * 0.0033, now * 0.00077));
+    uniforms.localToWorld = makeMatrix4RotationFromEuler(
+      uniforms.localToWorld,
+      new Euler(now * 0.001, now * 0.0033, now * 0.00077),
+    );
     canvasFramebuffer.renderBufferGeometry(program, uniforms, bufferGeometry, depthTestState);
   }
