+++ src/lib/math/Quaternion.ts
@@ -6,10 +6,7 @@
 //
 
 import { hashFloat4 } from "../core/hash";
-import { Euler, EulerOrder } from "./Euler";
 import { IPrimitive } from "./IPrimitive";
-import { Matrix4 } from "./Matrix4";
-import { Vector3 } from "./Vector3";
 
 export class Quaternion implements IPrimitive<Quaternion> {
   constructor(public x = 0, public y = 0, public z = 0, public w = 1) {}
@@ -145,139 +142,6 @@ export class Quaternion implements IPrimitive<Quaternion> {
       this.z *= l;
       this.w *= l;
     }
-
-    return this;
-  }
-
-  setFromEuler(euler: Euler): this {
-    const x = euler.x,
-      y = euler.y,
-      z = euler.z,
-      order = euler.order;
-
-    // eslint-disable-next-line max-len
-    // http://www.mathworks.com/matlabcentral/fileexchange/20696-function-to-convert-between-dcm-euler-angles-quaternions-and-euler-vectors/content/SpinCalc.m
-
-    const c1 = Math.cos(x / 2);
-    const c2 = Math.cos(y / 2);
-    const c3 = Math.cos(z / 2);
-
-    const s1 = Math.sin(x / 2);
-    const s2 = Math.sin(y / 2);
-    const s3 = Math.sin(z / 2);
-
-    switch (order) {
-      case EulerOrder.XYZ:
-        this.x = s1 * c2 * c3 + c1 * s2 * s3;
-        this.y = c1 * s2 * c3 - s1 * c2 * s3;
-        this.z = c1 * c2 * s3 + s1 * s2 * c3;
-        this.w = c1 * c2 * c3 - s1 * s2 * s3;
-        break;
-
-      case EulerOrder.YXZ:
-        this.x = s1 * c2 * c3 + c1 * s2 * s3;
-        this.y = c1 * s2 * c3 - s1 * c2 * s3;
-        this.z = c1 * c2 * s3 - s1 * s2 * c3;
-        this.w = c1 * c2 * c3 + s1 * s2 * s3;
-        break;
-
-      case EulerOrder.ZXY:
-        this.x = s1 * c2 * c3 - c1 * s2 * s3;
-        this.y = c1 * s2 * c3 + s1 * c2 * s3;
-        this.z = c1 * c2 * s3 + s1 * s2 * c3;
-        this.w = c1 * c2 * c3 - s1 * s2 * s3;
-        break;
-
-      case EulerOrder.ZYX:
-        this.x = s1 * c2 * c3 - c1 * s2 * s3;
-        this.y = c1 * s2 * c3 + s1 * c2 * s3;
-        this.z = c1 * c2 * s3 - s1 * s2 * c3;
-        this.w = c1 * c2 * c3 + s1 * s2 * s3;
-        break;
-
-      case EulerOrder.YZX:
-        this.x = s1 * c2 * c3 + c1 * s2 * s3;
-        this.y = c1 * s2 * c3 + s1 * c2 * s3;
-        this.z = c1 * c2 * s3 - s1 * s2 * c3;
-        this.w = c1 * c2 * c3 - s1 * s2 * s3;
-        break;
-
-      case EulerOrder.XZY:
-        this.x = s1 * c2 * c3 - c1 * s2 * s3;
-        this.y = c1 * s2 * c3 - s1 * c2 * s3;
-        this.z = c1 * c2 * s3 + s1 * s2 * c3;
-        this.w = c1 * c2 * c3 + s1 * s2 * s3;
-        break;
-    }
-
-    return this;
-  }
-
-  setFromRotationMatrix4(m: Matrix4): this {
-    // http://www.euclideanspace.com/maths/geometry/rotations/conversions/matrixToQuaternion/index.htm
-
-    // assumes the upper 3x3 of m is a pure rotation matrix (i.e, unscaled)
-
-    // TODO, allocate x, y, z, w and only set this.* at the end.
-
-    const te = m.elements,
-      m11 = te[0],
-      m12 = te[4],
-      m13 = te[8],
-      m21 = te[1],
-      m22 = te[5],
-      m23 = te[9],
-      m31 = te[2],
-      m32 = te[6],
-      m33 = te[10],
-      trace = m11 + m22 + m33;
-    let s;
-
-    if (trace > 0) {
-      s = 0.5 / Math.sqrt(trace + 1.0);
-
-      this.w = 0.25 / s;
-      this.x = (m32 - m23) * s;
-      this.y = (m13 - m31) * s;
-      this.z = (m21 - m12) * s;
-    } else if (m11 > m22 && m11 > m33) {
-      s = 2.0 * Math.sqrt(1.0 + m11 - m22 - m33);
-
-      this.w = (m32 - m23) / s;
-      this.x = 0.25 * s;
-      this.y = (m12 + m21) / s;
-      this.z = (m13 + m31) / s;
-    } else if (m22 > m33) {
-      s = 2.0 * Math.sqrt(1.0 + m22 - m11 - m33);
-
-      this.w = (m13 - m31) / s;
-      this.x = (m12 + m21) / s;
-      this.y = 0.25 * s;
-      this.z = (m23 + m32) / s;
-    } else {
-      s = 2.0 * Math.sqrt(1.0 + m33 - m11 - m22);
-
-      this.w = (m21 - m12) / s;
-      this.x = (m13 + m31) / s;
-      this.y = (m23 + m32) / s;
-      this.z = 0.25 * s;
-    }
-
-    return this;
-  }
-
-  setFromAxisAngle(axis: Vector3, angle: number): this {
-    // http://www.euclideanspace.com/maths/geometry/rotations/conversions/angleToQuaternion/index.htm
-
-    // assumes axis is normalized
-
-    const halfAngle = angle / 2,
-      s = Math.sin(halfAngle);
-
-    this.x = axis.x * s;
-    this.y = axis.y * s;
-    this.z = axis.z * s;
-    this.w = Math.cos(halfAngle);
 
     return this;
   }
