+++ src/lib/math/Vector3.ts
@@ -7,7 +7,6 @@
 
 import { hashFloat3 } from "../core/hash";
 import { IPrimitive } from "./IPrimitive";
-import { Matrix4 } from "./Matrix4";
 
 export class Vector3 implements IPrimitive<Vector3> {
   constructor(public x = 0, public y = 0, public z = 0) {}
@@ -144,23 +143,6 @@ export class Vector3 implements IPrimitive<Vector3> {
     this.x = ay * bz - az * by;
     this.y = az * bx - ax * bz;
     this.z = ax * by - ay * bx;
-
-    return this;
-  }
-
-  // TODO: think about moving this to a helper function -- it may allow for better code shaking...
-  // homogeneous coordinate projection
-  transformMatrix4(m: Matrix4): this {
-    const x = this.x,
-      y = this.y,
-      z = this.z;
-    const e = m.elements;
-
-    const w = 1 / (e[3] * x + e[7] * y + e[11] * z + e[15]);
-
-    this.x = (e[0] * x + e[4] * y + e[8] * z + e[12]) * w;
-    this.y = (e[1] * x + e[5] * y + e[9] * z + e[13]) * w;
-    this.z = (e[2] * x + e[6] * y + e[10] * z + e[14]) * w;
 
     return this;
   }
