+++ src/lib/nodes/cameras/OrthographicCamera.ts
@@ -6,6 +6,7 @@
 //
 
 import { Matrix4 } from "../../math/Matrix4";
+import { makeMatrix4Orthographic } from "../../math/Matrix4.Functions";
 import { Vector2 } from "../../math/Vector2";
 import { Camera } from "./Camera";
 
@@ -29,6 +30,6 @@ export class OrthographicCamera extends Camera {
     const top = -this.height * 0.5 + this.center.y;
     const bottom = top + this.height;
 
-    return new Matrix4().makeOrthographic(left, right, top, bottom, this.near, this.far);
+    return makeMatrix4Orthographic(new Matrix4(), left, right, top, bottom, this.near, this.far);
   }
 }
